requirements.txt
gradio
python-dotenv
google-generativeai
sentence-transformers
datasets
pdfplumber
torch
transformers